day06/src/requirements.txt
grpcio-tools
pydantic
pydantic-settings
psycopg2
sqlalchemy
